+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>me.noodles</groupId>
-    <artifactId>ss</artifactId>
+    <artifactId>ultimatefreeze</artifactId>
     <version>1.5</version>
     <packaging>jar</packaging>
