+++ src/main/java/me/noodles/ultimatefreeze/freezecommand/FreezeCommand.java
@@ -1,6 +1,6 @@
-package me.noodles.ss.freezecommand;
+package me.noodles.ultimatefreeze.freezecommand;
 
-import me.noodles.ss.Main;
+import me.noodles.ultimatefreeze.Main;
 import net.md_5.bungee.api.ChatColor;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
@@ -8,7 +8,6 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
-import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import java.util.ArrayList;
